Back/backend/requirements.txt
asgiref==3.5.2
asttokens==2.0.8
backcall==0.2.0
branca==0.5.0
certifi==2022.6.15.2
chardet==5.0.0
charset-normalizer==2.1.1
colorama==0.4.5
cycler==0.11.0
decorator==5.1.1
Django==4.1.1
djangorestframework==3.13.1
executing==1.0.0
folium==0.12.1.post1
fonttools==4.37.1
idna==3.3
ipython==8.4.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
numpy==1.22.2
packaging==21.3
pandas==1.4.4
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.31
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.11.2
six==1.16.0
sqlparse==0.4.2
stack-data==0.5.0
threadpoolctl==3.1.0
traitlets==5.3.0
tzdata==2022.2
urllib3==1.26.12
wcwidth==0.2.5
